msm: ep_pcie: Add D3 hot sleep feature
To save power and allow XO shutdown, turn off the pcie clocks
when device is in D3 hot and the link is in L1ss for a set period
of time (currently 1 second), based on the L1sub inactivity timer.
The CLKREQ sideband signal (gpio) is gated when the clocks are turned off,
and the wakeup capable clkreq gpio interrupt is enabled. Clocks
are turned back on when the clkreq interrupt is received.

Change-Id: Idff3c0db4a025ac3978fdb36c3c29532f87c1757
Signed-off-by: Siva Kumar Akkireddi <sivaa@codeaurora.org>